+++ pipeline/gog_pipeline/gog_extract.py
@@ -1,4 +1,6 @@
 """The extraction script for GOG"""
+from os import mkdir
+
 import re
 from time import sleep
 import json
@@ -10,40 +12,45 @@ from selenium.webdriver.chrome.service import Service
 from webdriver_manager.chrome import ChromeDriverManager
 
 
-def get_soup(url: str) -> BeautifulSoup:
+def init_driver():
+    """Sets up the selenium driver with proper service and options."""
+    tmp_dir = '/tmp/gc' 
+    mkdir(tmp_dir)
+    chrome_options = webdriver.ChromeOptions()
+    chrome_options.add_argument("--headless=new")
+    chrome_options.add_argument("--no-sandbox")
+    chrome_options.add_argument("--disable-dev-shm-usage")
+    chrome_options.add_argument("--disable-gpu")
+    chrome_options.add_argument("--disable-dev-tools")
+    chrome_options.add_argument("--no-zygote")
+    chrome_options.add_argument("--single-process")
+    chrome_options.add_argument(f"--user-data-dir={tmp_dir}")
+    chrome_options.add_argument(f"--data-path={tmp_dir}")
+    chrome_options.add_argument(f"--disk-cache-dir={tmp_dir}")
+    chrome_options.add_argument("--remote-debugging-pipe")
+    chrome_options.add_argument("--verbose")
+    chrome_options.add_argument("--log-path=/tmp")
+    chrome_options.binary_location = "/opt/chrome/chrome-linux64/chrome"
+
+    service = Service(
+        executable_path="/opt/chrome-driver/chromedriver-linux64/chromedriver",
+        service_log_path="/tmp/chromedriver.log"
+    )
+
+    driver = webdriver.Chrome(options=chrome_options, service=service)
+
+    return driver
+
+
+def get_soup(url: str, driver: webdriver.Chrome) -> BeautifulSoup:
     """Fetches the page content using Selenium and returns a BeautifulSoup object"""
-
-    options = webdriver.ChromeOptions()
-    options.add_argument("--headless")
-    driver = webdriver.Chrome(service=Service(
-        ChromeDriverManager().install()), options=options)
-
     driver.get(url)
-    sleep(3)
+    sleep(1)
     driver.execute_script("window.scrollTo(0, document.body.scrollHeight);") # Some resources only load when scrolling
-    sleep(3)
+    sleep(1)
     page_source = driver.page_source
-    driver.quit()
 
     return BeautifulSoup(page_source, "html.parser")
-
-
-def scrape_newest(url: str) -> list[dict]:
-    """
-    Scrapes all the newest games from GOG games
-    """
-    response = requests.get(url)
-
-    soup = BeautifulSoup(response.text, features='html.parser')
-    game_links = [link['href'] for link in soup.find_all('a', href=True)
-                  if re.match(r'https://www\.gog\.com/en/game/', link["href"])]
-
-
-    page_data_list = []
-    for link in game_links:
-        game_data = get_data(link)
-        page_data_list.append(game_data)
-    return page_data_list
 
 
 def fetch_title(soup: BeautifulSoup) -> str:
@@ -129,9 +136,9 @@ def fetch_age_rating(soup: BeautifulSoup) -> str:
         return match.group(1) if match else None
 
 
-def get_data(link: str) -> dict:
+def get_data(link: str, driver: webdriver) -> dict:
     """Gets the needed data from GOG website"""
-    soup = get_soup(link)
+    soup = get_soup(link, driver)
     data = {}
     data['title'] = fetch_title(soup)
     data['genres'] = fetch_genres(soup)
@@ -147,7 +154,36 @@ def get_data(link: str) -> dict:
     return data
 
 
+def scrape_newest(url: str, local:bool) -> list[dict]:
+    """
+    Scrapes all the newest games from GOG games
+    """
+    if local:
+        options = webdriver.ChromeOptions()
+        # options.add_argument("--headless")
+        driver = webdriver.Chrome(service=Service(
+            ChromeDriverManager().install()), options=options)
+    else:
+        driver = init_driver()
+
+    response = requests.get(url)
+
+    soup = BeautifulSoup(response.text, features='html.parser')
+    game_links = [link['href'] for link in soup.find_all('a', href=True)
+                  if re.match(r'https://www\.gog\.com/en/game/', link["href"])]
+
+
+    page_data_list = []
+    for link in game_links:
+        game_data = get_data(link, driver)
+        page_data_list.append(game_data)
+    driver.quit()
+    return page_data_list
+
+
 if __name__ == "__main__":
     data = scrape_newest(
-        'https://www.gog.com/en/games?releaseStatuses=new-arrival&order=desc:releaseDate&hideDLCs=true&releaseDateRange=2025,2025')
+        'https://www.gog.com/en/games?releaseStatuses=new-arrival&order=desc:releaseDate&hideDLCs=true&releaseDateRange=2025,2025',
+        True
+        )
     print(data)
